+++ packages/terra-framework-docs/src/terra-dev-site/doc/popup/example/PopupWithArrow.jsx
@@ -18,11 +18,15 @@ function PopupArrow() {
   const getButtonNode = () => buttonElement.current;
 
   const handleButtonClick = () => {
+    if (!open) {
+      document.querySelector('#root').setAttribute('inert', 'true');
+    }
     setOpen(true);
   };
 
   const handleRequestClose = () => {
     setOpen(false);
+    document.querySelector('#root').removeAttribute('inert');
     buttonElement.current.focus(); // Return focus to the button when the popup closes
   };
 
@@ -38,9 +42,6 @@ function PopupArrow() {
   useEffect(() => {
     if (open) {
       setFocusedIndex(0);
-      document.querySelector('#root').setAttribute('inert', 'true');
-    } else {
-      document.querySelector('#root').removeAttribute('inert');
     }
   }, [open]);