+++ packages/terra-framework-docs/src/terra-dev-site/doc/worklist-data-grid/Examples.2/RowSelection.jsx
@@ -1,5 +1,6 @@
 import React, { useState, useRef } from 'react';
 import WorklistDataGrid from 'terra-worklist-data-grid';
+import WorklistDataGridUtils from 'terra-worklist-data-grid/src/utils/WorklistDataGridUtils';
 
 const gridDataJSON = {
   cols: [
@@ -218,6 +219,13 @@ const RowSelection = () => {
     setHasSelectableRows(event.target.checked);
   };
 
+  const onColumnSelect = (columnId) => {
+    if (columnId === WorklistDataGridUtils.ROW_SELECTION_COLUMN.id) {
+      // eslint-disable-next-line no-alert
+      alert('Row Selection Header Clicked');
+    }
+  };
+
   return (
     <React.Fragment>
       <div>
@@ -272,6 +280,7 @@ const RowSelection = () => {
         onDisableSelectableRows={() => {
           disableSelectableRows();
         }}
+        onColumnSelect={onColumnSelect}
         onEnableRowSelection={() => {
           if (!rowSelectionModeRef.current.checked) {
             rowSelectionModeRef.current.checked = true;