+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,11 +5,13 @@
 * Changed
   * Updated toggle examples for `terra-collapsible-menu-view` to include selectable toggles and toggles within a collapsed menu.
 
+* Added
+  * Updated tests for `terra-tabs` drag and drop in tab-menu.
+
 ## 1.31.0 - (August 4, 2023)
 
 * Added
   * Added examples and tests for `terra-worklist-data-grid` to interact with cells containing focusable elements.
-  * Updated tests for `terra-tabs` drag and drop in tab-menu.
 
 ## 1.30.0 - (August 2, 2023)