+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Updated
-  * Examples with inert attribute to ensure accessibility when arrow keys are used for `terra-popup`.
+  * Updated terra-popup examples with inert attribute to demonstrate focus trap when using virtual cursor of Jaws for navigation within popup.
 
 ## 1.31.0 - (August 4, 2023)