+++ packages/terra-framework-docs/CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
 ## Unreleased
+
+* Updated
+  * Updated terra-popup examples with inert attribute to demonstrate focus trap when using virtual cursor of Jaws for navigation within popup.
+
+* Changed
+  * Updated toggle examples for `terra-collapsible-menu-view` to include selectable toggles and toggles within a collapsed menu.
+  * Updated row selection example for `terra-worklist-data-grid` to show ability to select the row selection column header.
+
+* Changed
+ * Locked `terra-arrange` to `3.51.0` temporarily.
+
+## 1.31.0 - (August 4, 2023)
+
 * Added
   * Added examples and tests for `terra-worklist-data-grid` to interact with cells containing focusable elements.